+++ tests/Value/LocalityTest.php
@@ -8,7 +8,8 @@ use DigipolisGent\Flanders\BasicRegisters\Value\GeographicalName;
 use DigipolisGent\Flanders\BasicRegisters\Value\GeographicalNames;
 use DigipolisGent\Flanders\BasicRegisters\Value\LanguageCode;
 use DigipolisGent\Flanders\BasicRegisters\Value\Locality;
-use DigipolisGent\Flanders\BasicRegisters\Value\ObjectId;
+use DigipolisGent\Flanders\BasicRegisters\Value\LocalityId;
+use DigipolisGent\Flanders\BasicRegisters\Value\PostInfoId;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -23,17 +24,19 @@ class LocalityTest extends TestCase
      *
      * @test
      */
-    public function createdFromObjectIdAndGeographicalNames(): void
+    public function valueIsCreatedFromItsDetails(): void
     {
-        $objectId = new ObjectId(123);
+        $localityId = new LocalityId(123);
         $geographicalNames = new GeographicalNames(
             new GeographicalName(new LanguageCode('NL'), 'Foo Nl')
         );
+        $postInfoId = new PostInfoId(9000);
 
-        $locality = new Locality($objectId, $geographicalNames);
+        $locality = new Locality($localityId, $geographicalNames, $postInfoId);
 
-        $this->assertSame($objectId, $locality->objectId());
+        $this->assertSame($localityId, $locality->localityId());
         $this->assertSame($geographicalNames, $locality->geographicalNames());
+        $this->assertSame($postInfoId, $locality->postInfoId());
     }
 
     /**
@@ -43,15 +46,34 @@ class LocalityTest extends TestCase
      */
     public function nameIsExtractedFromGeographicalNames(): void
     {
-        $objectId = new ObjectId(123);
+        $localityId = new LocalityId(123);
         $geographicalNames = new GeographicalNames(
             new GeographicalName(new LanguageCode('EN'), 'Foo EN'),
             new GeographicalName(new LanguageCode('NL'), 'Foo Nl')
         );
+        $postInfoId = new PostInfoId(9000);
 
-        $locality = new Locality($objectId, $geographicalNames);
+        $locality = new Locality($localityId, $geographicalNames, $postInfoId);
 
         $this->assertSame($geographicalNames->name(), $locality->name());
+    }
+
+    /**
+     * Postal code is extracted from the postInfoId.
+     *
+     * @test
+     */
+    public function postalCodeIsExtractedFromPostInfoId(): void
+    {
+        $localityId = new LocalityId(123);
+        $geographicalNames = new GeographicalNames(
+            new GeographicalName(new LanguageCode('NL'), 'Foo Nl')
+        );
+        $postInfoId = new PostInfoId(9000);
+
+        $locality = new Locality($localityId, $geographicalNames, $postInfoId);
+
+        $this->assertSame($postInfoId->value(), $locality->postalCode());
     }
 
     /**
@@ -59,17 +81,18 @@ class LocalityTest extends TestCase
      *
      * @test
      */
-    public function notSameIfObjectIdIsDifferent(): void
+    public function notSameIfLocalityIdIsDifferent(): void
     {
+        $localityId = new LocalityId(123);
         $geographicalNames = new GeographicalNames(
             new GeographicalName(new LanguageCode('EN'), 'Foo EN')
         );
+        $postInfoId = new PostInfoId(9000);
 
-        $objectId = new ObjectId(123);
-        $locality = new Locality($objectId, $geographicalNames);
+        $locality = new Locality($localityId, $geographicalNames, $postInfoId);
 
-        $otherObjectId = new ObjectId(456);
-        $otherLocality = new Locality($otherObjectId, $geographicalNames);
+        $otherLocalityId = new LocalityId(456);
+        $otherLocality = new Locality($otherLocalityId, $geographicalNames, $postInfoId);
 
         $this->assertFalse($locality->sameValueAs($otherLocality));
     }
@@ -81,35 +104,58 @@ class LocalityTest extends TestCase
      */
     public function notSameIfGeographicalNamesAreDifferent(): void
     {
-        $objectId = new ObjectId(123);
-
+        $localityId = new LocalityId(123);
         $geographicalNames = new GeographicalNames(
             new GeographicalName(new LanguageCode('EN'), 'Foo EN')
         );
-        $locality = new Locality($objectId, $geographicalNames);
+        $postInfoId = new PostInfoId(9000);
+
+        $locality = new Locality($localityId, $geographicalNames, $postInfoId);
 
         $otherGeographicalNames = new GeographicalNames(
             new GeographicalName(new LanguageCode('NL'), 'Foo NL')
         );
-        $otherLocality = new Locality($objectId, $otherGeographicalNames);
+        $otherLocality = new Locality($localityId, $otherGeographicalNames, $postInfoId);
 
         $this->assertFalse($locality->sameValueAs($otherLocality));
     }
 
     /**
-     * Same values if object id and geographical names are identical.
+     * Not the same value if the post info ids are different.
      *
      * @test
      */
-    public function sameIfObjectIdAndGeographicalNamesAreIdentical(): void
+    public function notSameIfPostInfoIdsAreDifferent(): void
     {
-        $objectId = new ObjectId(123);
+        $localityId = new LocalityId(123);
         $geographicalNames = new GeographicalNames(
             new GeographicalName(new LanguageCode('EN'), 'Foo EN')
         );
+        $postInfoId = new PostInfoId(9000);
 
-        $locality = new Locality($objectId, $geographicalNames);
-        $sameLocality = new Locality($objectId, $geographicalNames);
+        $locality = new Locality($localityId, $geographicalNames, $postInfoId);
+
+        $otherPostInfoId = new PostInfoId(9123);
+        $otherLocality = new Locality($localityId, $geographicalNames, $otherPostInfoId);
+
+        $this->assertFalse($locality->sameValueAs($otherLocality));
+    }
+
+    /**
+     * Same values if all details are identical.
+     *
+     * @test
+     */
+    public function sameIfAllDetailsAreIdentical(): void
+    {
+        $localityId = new LocalityId(123);
+        $geographicalNames = new GeographicalNames(
+            new GeographicalName(new LanguageCode('EN'), 'Foo EN')
+        );
+        $postInfoId = new PostInfoId(9000);
+
+        $locality = new Locality($localityId, $geographicalNames, $postInfoId);
+        $sameLocality = new Locality($localityId, $geographicalNames, $postInfoId);
 
         $this->assertTrue($locality->sameValueAs($sameLocality));
     }
@@ -121,11 +167,13 @@ class LocalityTest extends TestCase
      */
     public function castToStringReturnsNameMethodValue(): void
     {
-        $objectId = new ObjectId(123);
+        $localityId = new LocalityId(123);
         $geographicalNames = new GeographicalNames(
             new GeographicalName(new LanguageCode('EN'), 'Foo EN')
         );
-        $locality = new Locality($objectId, $geographicalNames);
+        $postInfoId = new PostInfoId(9000);
+
+        $locality = new Locality($localityId, $geographicalNames, $postInfoId);
 
         $this->assertSame($locality->name(), (string) $locality);
     }