+++ CHANGELOG.md
@@ -9,5 +9,7 @@ All Notable changes to `digipolisgent/flanders-basicregisters` package.
 * Added the Locality value object.
 * Added the StreetName value object.
 * Added the address point values.
+* Added the addresses collection.
+* Added the address detail.
 
 [Unreleased]: https://github.com/digipolisgent/php_package_dg-flanders-basicregisters/compare/master...develop